+++ lib/record-serialization.js
@@ -11,16 +11,16 @@ module.exports = jsrs;
 // one of our priorities to optimize it.
 const USE_NATIVE_SERIALIZER = false;
 
-let jsrsNative;
+let jstpNative;
 
 try {
-  jsrsNative = require('../build/Release/jsrs');
+  jstpNative = require('../build/Release/jstp');
 } catch (e) {
   try {
-    jsrsNative = require('../build/Debug/jsrs');
+    jstpNative = require('../build/Debug/jstp');
   } catch (e) {
     console.warn(
-      'JSTP Record Serialization native module is not built. ' +
+      'JSTP native addon is not built. ' +
       'Run `npm install` in order to build it, otherwise you will get ' +
       'poor server performance under load.'
     );
@@ -28,8 +28,8 @@ try {
   }
 }
 
-if (jsrsNative) {
-  Object.assign(jsrs, jsrsNative);
+if (jstpNative) {
+  Object.assign(jsrs, jstpNative);
   if (!USE_NATIVE_SERIALIZER) {
     jsrs.stringify = serializerFactory.createSerializer();
   }
