+++ src/jsrs_parser.cc
@@ -1,0 +1,654 @@
+// Copyright (c) 2016-2017 JSTP project authors. Use of this source code is
+// governed by the MIT license that can be found in the LICENSE file.
+
+#include "jsrs_parser.h"
+
+#include <cctype>
+#include <cstddef>
+#include <cstdlib>
+#include <cstring>
+
+#include <vector>
+
+#include "unicode_utils.h"
+
+using std::ptrdiff_t;
+using std::size_t;
+
+using v8::Array;
+using v8::Exception;
+using v8::False;
+using v8::Integer;
+using v8::Isolate;
+using v8::Local;
+using v8::Maybe;
+using v8::NewStringType;
+using v8::Null;
+using v8::Number;
+using v8::Object;
+using v8::String;
+using v8::True;
+using v8::Undefined;
+using v8::Value;
+
+using jstp::unicode_utils::CodePointToUtf8;
+using jstp::unicode_utils::IsWhiteSpaceCharacter;
+using jstp::unicode_utils::IsLineTerminatorSequence;
+
+namespace jstp {
+
+namespace parser {
+
+// Enumeration of supported JavaScript types used for deserialization
+// function selection.
+enum Type {
+  kUndefined = 0, kNull, kBool, kNumber, kString, kArray, kObject, kDate
+};
+
+// Parses the type of the serialized JavaScript value at the position `begin`
+// and before `end`. Returns true if it was able to detect the type, false
+// otherwise.
+static bool GetType(const char* begin, const char* end, Type* type);
+
+// The table of parsing functions indexed with the values of the Type
+// enumeration.
+static Local<Value> (*kParseFunctions[])(Isolate*,
+                                         const char*,
+                                         const char*,
+                                         size_t*) = {
+  &internal::ParseUndefined,
+  &internal::ParseNull,
+  &internal::ParseBool,
+  &internal::ParseNumber,
+  &internal::ParseString,
+  &internal::ParseArray,
+  &internal::ParseObject
+};
+
+Local<Value> Parse(Isolate* isolate, const String::Utf8Value& in) {
+  size_t size;
+  const char* to_parse =
+      internal::PrepareString(isolate, *in, in.length(), &size);
+  if (!to_parse) {
+    return Undefined(isolate);
+  }
+
+  Type type;
+  if (!GetType(to_parse, to_parse + size, &type)) {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Invalid type")));
+    return Undefined(isolate);
+  }
+
+  size_t parsed_size = 0;
+  Local<Value> result =
+      (kParseFunctions[type])(isolate, to_parse, to_parse + size, &parsed_size);
+  if (size != parsed_size) {
+    isolate->ThrowException(Exception::SyntaxError(
+        String::NewFromUtf8(isolate, "Invalid format")));
+    return Undefined(isolate);
+  }
+
+  delete[] to_parse;
+  return result;
+}
+
+static bool GetType(const char* begin, const char* end, Type* type) {
+  bool result = true;
+  switch (*begin) {
+    case ',':
+    case ']':
+      *type = Type::kUndefined;
+      break;
+    case '{':
+      *type = Type::kObject;
+      break;
+    case '[':
+      *type = Type::kArray;
+      break;
+    case '\"':
+    case '\'':
+      *type = Type::kString;
+      break;
+    case 't':
+    case 'f':
+      *type = Type::kBool;
+      break;
+    case 'n':
+      *type = Type::kNull;
+      if (begin + 4 <= end) {
+        result = (strncmp(begin, "null", 4) == 0);
+      }
+      break;
+    case 'u':
+      *type = Type::kUndefined;
+      if (begin + 9 <= end) {
+        result = (strncmp(begin, "undefined", 9) == 0);
+      }
+      break;
+    default:
+      result = false;
+      if (isdigit(*begin) || *begin == '.' || *begin == '+' || *begin == '-') {
+        *type = Type::kNumber;
+        result = true;
+      }
+  }
+  return result;
+}
+
+namespace internal {
+
+const char* PrepareString(Isolate*    isolate,
+                          const char* str,
+                          size_t      length,
+                          size_t*     new_length) {
+  char* result = new char[length + 1];
+  bool string_mode = false;
+  enum { kDisabled = 0, kOneline, kMultiline } comment_mode = kDisabled;
+  size_t j = 0;
+  size_t size = 0;
+
+  for (size_t i = 0; i < length; i++) {
+    if ((comment_mode == kDisabled) &&
+        (str[i] == '\"' || str[i] == '\'') &&
+        (i == 0 || str[i - 1] != '\\')) {
+      string_mode = !string_mode;
+    }
+
+    if (!string_mode) {
+      if (comment_mode == kDisabled && str[i] == '/') {
+        switch (str[i + 1]) {
+          case '/':
+            comment_mode = kOneline;
+            break;
+          case '*':
+            comment_mode = kMultiline;
+            break;
+        }
+      }
+
+      if (comment_mode == kDisabled) {
+        if (IsWhiteSpaceCharacter(str + i, &size) ||
+            IsLineTerminatorSequence(str + i, &size)) {
+          i += size - 1;
+        } else {
+          result[j++] = str[i];
+        }
+      }
+
+      if ((comment_mode == kOneline &&
+            IsLineTerminatorSequence(str + i, &size)) ||
+          (comment_mode == kMultiline &&
+            str[i - 1] == '*' && str[i] == '/')) {
+        comment_mode = kDisabled;
+      }
+    } else if (str[i] == '\\' && IsLineTerminatorSequence(str + i + 1, &size)) {
+      i += size;
+    } else if (IsLineTerminatorSequence(str + i, &size)) {
+      isolate->ThrowException(Exception::SyntaxError(
+          String::NewFromUtf8(isolate, "Unexpected line end in string")));
+      delete[] result;
+      return nullptr;
+    } else {
+      result[j++] = str[i];
+    }
+  }
+
+  result[j] = '\0';
+  *new_length = j;
+
+  return result;
+}
+
+Local<Value> ParseUndefined(Isolate*    isolate,
+                            const char* begin,
+                            const char* end,
+                            size_t*     size) {
+  if (*begin == ',' || *begin == ']') {
+    *size = 0;
+  } else if (*begin == 'u') {
+    *size = 9;
+  } else {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Invalid format of undefined value")));
+  }
+  return Undefined(isolate);
+}
+
+Local<Value> ParseNull(Isolate*    isolate,
+                       const char* begin,
+                       const char* end,
+                       size_t*     size) {
+  *size = 4;
+  return Null(isolate);
+}
+
+Local<Value> ParseBool(Isolate*    isolate,
+                       const char* begin,
+                       const char* end,
+                       size_t*     size) {
+  Local<Value> result;
+  if (begin + 4 <= end && strncmp(begin, "true", 4) == 0) {
+    result = True(isolate);
+    *size = 4;
+  } else if (begin + 5 <= end && strncmp(begin, "false", 5) == 0) {
+    result = False(isolate);
+    *size = 5;
+  } else {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate,
+            "Invalid format: expected boolean")));
+    result = Undefined(isolate);
+  }
+  return result;
+}
+
+// Checks if a character is an octal digit.
+inline bool IsOctalDigit(char character) {
+  return character >= '0' && character <= '7';
+}
+
+Local<Value> ParseNumber(Isolate*    isolate,
+                         const char* begin,
+                         const char* end,
+                         size_t*     size) {
+  bool negate_result = false;
+  const char* number_start = begin;
+
+  if (*begin == '+' || *begin == '-') {
+    negate_result = *begin == '-';
+    number_start++;
+  }
+
+  int base = 10;
+
+  if (*number_start == '0') {
+    number_start++;
+
+    if (IsOctalDigit(*number_start)) {
+      return isolate->ThrowException(Exception::SyntaxError(
+          String::NewFromUtf8(isolate, "Use new octal literal syntax")));
+    } else if (*number_start == 'b') {
+      base = 2;
+      number_start++;
+    } else if (*number_start == 'o') {
+      base = 8;
+      number_start++;
+    } else if (*number_start == 'x') {
+      base = 16;
+      number_start++;
+    } else {
+      number_start--;
+    }
+  }
+
+  if (base == 10) {
+    return ParseDecimalNumber(isolate, begin, end, size);
+  } else {
+    auto value = ParseIntegerNumber(isolate, number_start, end, size,
+                                    base, negate_result);
+    auto offset = static_cast<size_t>(number_start - begin);
+    *size += offset;
+    return value;
+  }
+}
+
+Local<Value> ParseDecimalNumber(Isolate*    isolate,
+                                const char* begin,
+                                const char* end,
+                                size_t*     size) {
+  auto result = Number::New(isolate, atof(begin));
+  *size = end - begin;
+  size_t i = 0;
+  while (begin[i] != ',' &&
+         begin[i] != '}' &&
+         begin[i] != ']' &&
+         i < *size) {
+    i++;
+  }
+  *size = i;
+  return result;
+}
+
+Local<Value> ParseIntegerNumber(Isolate*    isolate,
+                                const char* begin,
+                                const char* end,
+                                size_t*     size,
+                                int         base,
+                                bool        negate_result) {
+  char* number_end;
+  int32_t value = strtol(begin, &number_end, base);
+  if (negate_result) {
+    value = -value;
+  }
+  *size = static_cast<size_t>(number_end - begin);
+  return Integer::New(isolate, value);
+}
+
+static char* GetControlChar(Isolate*    isolate,
+                            const char* str,
+                            size_t*     res_len,
+                            size_t*     size);
+
+Local<Value> ParseString(Isolate*    isolate,
+                         const char* begin,
+                         const char* end,
+                         size_t*     size) {
+  *size = end - begin;
+  char* result = new char[*size + 1];
+  memset(result, 0, *size + 1);
+
+  enum { kApostrophe = 0, kQMarks} string_mode = (*begin == '\'') ?
+                                                 kApostrophe :
+                                                 kQMarks;
+  bool is_ended = false;
+  size_t res_index = 0;
+  size_t out_offset, in_offset;
+
+  for (size_t i = 1; i < *size; i++) {
+    if ((string_mode == kQMarks     && begin[i] == '\"') ||
+        (string_mode == kApostrophe && begin[i] == '\'')) {
+      is_ended = true;
+      *size = i + 1;
+      result[res_index] = '\0';
+      break;
+    }
+
+    if (begin[i] == '\\') {
+      char* symb =
+          GetControlChar(isolate, begin + ++i, &out_offset, &in_offset);
+      if (!symb) {
+        return String::Empty(isolate);
+      }
+      strncpy(result + res_index, symb, out_offset);
+      delete[] symb;
+      i += in_offset - 1;
+      res_index += out_offset;
+    } else {
+      result[res_index++] = begin[i];
+    }
+  }
+
+  if (!is_ended) {
+    isolate->ThrowException(Exception::SyntaxError(
+        String::NewFromUtf8(isolate, "Error while parsing string")));
+    return String::Empty(isolate);
+  }
+
+  Local<String> result_str = String::NewFromUtf8(isolate, result,
+      NewStringType::kNormal, static_cast<int>(res_index)).ToLocalChecked();
+  delete[] result;
+  return result_str;
+}
+
+static unsigned int ReadHexNumber(const char* str, size_t len, bool* ok);
+
+// Parses a part of a JavaScript string representation after the backslash
+// character (i.e., an escape sequence without \) into an unescaped control
+// character.
+static char* GetControlChar(Isolate*    isolate,
+                            const char* str,
+                            size_t*     res_len,
+                            size_t*     size) {
+  char* result = new char[5];
+  *size = 1;
+  *res_len = 1;
+  bool ok;
+  switch (str[0]) {
+    case 'b': *result = '\b'; break;
+    case 'f': *result = '\f'; break;
+    case 'n': *result = '\n'; break;
+    case 'r': *result = '\r'; break;
+    case 't': *result = '\t'; break;
+    case 'v': *result = '\v'; break;
+    case '0': *result = '\0'; break;
+
+    case 'x': {
+      *result = ReadHexNumber(str + 1, 2, &ok);
+      if (!ok) {
+        delete[] result;
+        isolate->ThrowException(Exception::SyntaxError(
+            String::NewFromUtf8(isolate,
+                "Invalid hexadecimal escape sequence")));
+        return nullptr;
+      }
+      *size = 3;
+      break;
+    }
+
+    case 'u': {
+      unsigned int symb_code;
+      if (isxdigit(str[1])) {
+        symb_code = ReadHexNumber(str + 1, 4, &ok);
+        *size = 5;
+      } else if (str[1] == '{') {
+        size_t hex_size;  // maximal hex is 10FFFF
+        for (hex_size = 1;
+             str[hex_size + 2] != '}' && hex_size <= 6;
+             hex_size++) {
+          if (str[hex_size + 2] == '\0') {
+            delete[] result;
+            isolate->ThrowException(Exception::SyntaxError(
+                String::NewFromUtf8(isolate,
+                    "Invalid Unicode code point escape")));
+            return nullptr;
+          }
+        }
+        symb_code = ReadHexNumber(str + 2, hex_size, &ok);
+        *size = hex_size + 3;
+      } else {
+        ok = false;
+      }
+
+      if (!ok) {
+        delete[] result;
+        isolate->ThrowException(Exception::SyntaxError(
+            String::NewFromUtf8(isolate,
+                "Invalid Unicode escape sequence")));
+        return nullptr;
+      }
+      char* unicode_symbol = CodePointToUtf8(symb_code, res_len);
+      delete[] result;
+      return unicode_symbol;
+    }
+
+    default:
+      *result = str[0];
+  }
+
+  return result;
+}
+
+// Parses a hexadecimal number into unsigned int. Whether the parsing
+// was successful is determined by the value of `ok`.
+static unsigned int ReadHexNumber(const char* str, size_t len, bool* ok) {
+  char t[6];
+  char* end;
+  strncpy(t, str, len);
+  t[len] = '\0';
+  unsigned int result = strtol(t, &end, 16);
+  if (end - t != static_cast<ptrdiff_t>(len)) {
+    *ok = false;
+  } else {
+    *ok = true;
+  }
+  return result;
+}
+
+Local<String> ParseKeyInObject(Isolate*    isolate,
+                               const char* begin,
+                               const char* end,
+                               size_t*     size) {
+  *size = end - begin;
+  Local<String> result;
+  if (begin[0] == '\'' || begin[0] == '"') {
+    Type current_type;
+    bool valid = GetType(begin, end, &current_type);
+    if (valid && current_type == Type::kString) {
+      size_t offset;
+      result = ParseString(isolate, begin, end, &offset).As<String>();
+      *size = offset;
+      return result;
+    } else {
+      isolate->ThrowException(Exception::SyntaxError(
+          String::NewFromUtf8(isolate,
+              "Invalid format in object: key is invalid string")));
+      return Local<String>();
+    }
+  } else {
+    size_t current_length = 0;
+    for (size_t i = 0; i < *size; i++) {
+      if (begin[i] == ':') {
+        if (current_length != 0) {
+          result = String::NewFromUtf8(isolate, begin,
+                                       NewStringType::kInternalized,
+                                       static_cast<int>(current_length))
+                                           .ToLocalChecked();
+          break;
+        } else {
+          isolate->ThrowException(Exception::SyntaxError(
+              String::NewFromUtf8(isolate, "Unexpected token :")));
+          return Local<String>();
+        }
+      } else if (begin[i] == '_' || (i != 0 ?
+                                     isalnum(begin[i]) :
+                                     isalpha(begin[i]))) {
+        current_length++;
+      } else {
+        isolate->ThrowException(Exception::SyntaxError(
+            String::NewFromUtf8(isolate,
+                "Invalid format in object: key has invalid type")));
+        return Local<String>();
+      }
+    }
+    *size = current_length;
+    return result;
+  }
+}
+
+Local<Value> ParseValueInObject(Isolate*    isolate,
+                                const char* begin,
+                                const char* end,
+                                size_t*     size) {
+  Local<Value> value;
+  Type current_type;
+  bool valid = GetType(begin, end, &current_type);
+  if (valid) {
+    value = (kParseFunctions[current_type])(isolate, begin, end, size);
+    return value;
+  } else {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Invalid type in object")));
+    return Object::New(isolate);
+  }
+}
+
+Local<Value> ParseObject(Isolate*    isolate,
+                         const char* begin,
+                         const char* end,
+                         size_t*     size) {
+  bool key_mode = true;
+  *size = end - begin;
+  Local<String> current_key;
+  Local<Value> current_value;
+  size_t current_length = 0;
+  auto result = Object::New(isolate);
+
+  for (size_t i = 1; i < *size; i++) {
+    if (key_mode) {
+      if (begin[i] == '}') {
+        if (begin[i - 1] != ',') {  // In case of empty object
+          *size = 2;
+        } else {                    // In case of trailing comma
+          *size = i + 1;
+        }
+        break;
+      }
+      current_key = ParseKeyInObject(isolate,
+                                     begin + i,
+                                     end,
+                                     &current_length);
+      i += current_length;
+    } else {
+      current_value = ParseValueInObject(isolate,
+                                         begin + i,
+                                         end,
+                                         &current_length);
+      if (!current_value->IsUndefined()) {
+        Maybe<bool> is_ok = result->Set(isolate->GetCurrentContext(),
+                                        current_key,
+                                        current_value);
+        if (is_ok.IsNothing()) {
+          isolate->ThrowException(
+              Exception::Error(String::NewFromUtf8(isolate,
+                  "Cannot add property to object")));
+        }
+      }
+      i += current_length;
+      if (begin[i] != ',' && begin[i] != '}') {
+        isolate->ThrowException(Exception::SyntaxError(
+            String::NewFromUtf8(isolate,
+                "Invalid format in object")));
+        return Object::New(isolate);
+      } else if (begin[i] == '}') {
+        *size = i + 1;
+        break;
+      }
+    }
+    key_mode = !key_mode;
+  }
+  return result;
+}
+
+Local<Value> ParseArray(Isolate*    isolate,
+                        const char* begin,
+                        const char* end,
+                        size_t*     size) {
+  auto array = Array::New(isolate);
+  size_t current_length = 0;
+  *size = end - begin;
+  if (*begin == '[' && *(begin + 1) == ']') {  // In case of empty array
+    *size = 2;
+    return array;
+  }
+
+  size_t current_element = 0;
+  for (size_t i = 1; i < *size; i++) {
+    Type current_type;
+    bool valid = GetType(begin + i, end, &current_type);
+    if (valid) {
+      auto t = kParseFunctions[current_type](isolate,
+                                             begin + i,
+                                             end,
+                                             &current_length);
+      if (!(current_type == Type::kUndefined && begin[i] == ']')) {
+        array->Set(static_cast<uint32_t>(current_element++), t);
+      }
+
+      i += current_length;
+      current_length = 0;
+
+      if (begin[i] != ',' && begin[i] != ']') {
+        isolate->ThrowException(Exception::SyntaxError(
+            String::NewFromUtf8(isolate,
+                "Invalid format in array: missed comma")));
+        return Array::New(isolate);
+      } else if (begin[i] == ']') {
+        *size = i + 1;
+        break;
+      }
+    } else {
+      isolate->ThrowException(Exception::TypeError(
+          String::NewFromUtf8(isolate, "Invalid type in array")));
+      return Array::New(isolate);
+    }
+  }
+
+  return array;
+}
+
+}  // namespace internal
+
+}  // namespace parser
+
+}  // namespace jstp
