+++ src/node_bindings.cc
@@ -1,0 +1,96 @@
+// Copyright (c) 2016-2017 JSTP project authors. Use of this source code is
+// governed by the MIT license that can be found in the LICENSE file.
+
+#include <node.h>
+#include <v8.h>
+
+#include "jsrs_parser.h"
+#include "jsrs_serializer.h"
+#include "packet_parser.h"
+
+using v8::Array;
+using v8::Exception;
+using v8::FunctionCallbackInfo;
+using v8::HandleScope;
+using v8::Isolate;
+using v8::Local;
+using v8::Object;
+using v8::String;
+using v8::Value;
+
+namespace jstp {
+
+namespace bindings {
+
+void Stringify(const FunctionCallbackInfo<Value>& args) {
+  Isolate* isolate = args.GetIsolate();
+
+  if (args.Length() != 1) {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Wrong number of arguments")));
+    return;
+  }
+
+  HandleScope scope(isolate);
+
+  Local<String> result = jstp::serializer::Stringify(isolate, args[0]);
+  args.GetReturnValue().Set(result);
+}
+
+void Parse(const FunctionCallbackInfo<Value>& args) {
+  Isolate* isolate = args.GetIsolate();
+
+  if (args.Length() != 1) {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Wrong number of arguments")));
+    return;
+  }
+  if (!args[0]->IsString() && !args[0]->IsUint8Array()) {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Wrong argument type")));
+    return;
+  }
+
+  HandleScope scope(isolate);
+
+  String::Utf8Value str(args[0]->ToString());
+
+  Local<v8::Value> result = jstp::parser::Parse(isolate, str);
+  args.GetReturnValue().Set(result);
+}
+
+void ParseNetworkPackets(const FunctionCallbackInfo<Value>& args) {
+  Isolate* isolate = args.GetIsolate();
+
+  if (args.Length() != 2) {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Wrong number of arguments")));
+    return;
+  }
+
+  if (!args[0]->IsString() || !args[1]->IsArray()) {
+    isolate->ThrowException(Exception::TypeError(
+        String::NewFromUtf8(isolate, "Wrong argument type")));
+    return;
+  }
+
+  HandleScope scope(isolate);
+
+  String::Utf8Value str(args[0]->ToString());
+  auto array = Local<Array>::Cast(args[1]);
+  auto result = jstp::packet_parser::ParseNetworkPackets(isolate, str, array);
+
+  args.GetReturnValue().Set(result);
+}
+
+void Init(Local<Object> target) {
+  NODE_SET_METHOD(target, "stringify", Stringify);
+  NODE_SET_METHOD(target, "parse", Parse);
+  NODE_SET_METHOD(target, "parseNetworkPackets", ParseNetworkPackets);
+}
+
+NODE_MODULE(jsrs, Init);
+
+}  // namespace bindings
+
+}  // namespace jstp
